Limit SWA KV metadata to unified cache
